==> scripts/03_Post_COMPASS_Plots.Rmd <==
---
title: "Post-COMPASS Plots"
author: "Jolie Phan"
output: 
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
date: "version `r format(Sys.time(), '%B %d, %Y')`"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

# Load libraries

```{r, message=F}
library(here)
library(tidyverse)
library(COMPASS)
library(ggplot2)
library(data.table)
library(broom)
library(psych)
library(corrplot)
library(ggpubr)
library(knitr)
library(kableExtra)
library(cowplot)
library(purrr)
library(flowWorkspace)
library(GGally)
library(patchwork)
library(ComplexHeatmap)
library(ggbeeswarm)
library(ggh4x)
source(here::here("scripts/Helper_Functions.R"))
```

# Set font

```{r}
save_output <- TRUE
# Arial font setup. Downloaded afms from https://github.com/microsoft/microsoft-r-open/tree/ec3fd89e5fb5794bd8149905c134ad801bb61800
Arial <- Type1Font(family = "Arial",
                   metrics = c(here::here("data/Arial_afm/ArialMT.afm"),
                               here::here("data/Arial_afm/ArialMT-Bold.afm"),
                               here::here("data/Arial_afm/ArialMT-Italic.afm"),
                               here::here("data/Arial_afm/ArialMT-BoldItalic.afm")))
windowsFonts("Arial" = windowsFont("Arial"))
pdfFonts(Arial = Arial)

# Naive PRE is light blue, Naive POST is dark blue. "#54A8E1" "#196293"
# Conv PRE is light red, Conv POST is dark red "#EB4D4F" "#9D1213"
```

# Read in data

```{r read in data}
# Next 4 lines from 02_Run_COMPASS.R
stims_for_compass_runs <- c("S1", "S2", "NCAP")
parent_nodes_for_compass_runs <- c("CD4+", "CD8+")
stims_for_compass_runs_rep <- rep(stims_for_compass_runs, each = length(parent_nodes_for_compass_runs))
parent_nodes_for_compass_runs_rep <- rep(parent_nodes_for_compass_runs, times = length(stims_for_compass_runs))
crList <- purrr::pmap(.l = list(parent_nodes_for_compass_runs_rep,
                                stims_for_compass_runs_rep),
                      .f = function(parent, currentStim) {
                        currentStimForFile <- gsub(" ", "_", currentStim)
                        crPath <- here::here(sprintf("out/CompassOutput/%s/%s/COMPASSResult_%s_%s.rds", parent, currentStimForFile, parent, currentStimForFile))
                        readRDS(crPath)
                      }) %>% 
  set_names(gsub("+", "", gsub(" ", "_", paste0(parent_nodes_for_compass_runs_rep, "_", stims_for_compass_runs_rep)), fixed=TRUE))
names(crList)
CD4RunNames <- c("CD4_S1", "CD4_S2", "CD4_NCAP")
CD8RunNames <- c("CD8_S1", "CD8_S2", "CD8_NCAP")

gsPath <- here::here("out/GatingSets/HAARVIVAC_GatingSet")
gs <- load_gs(gsPath)

if(!dir.exists(here::here("processed_data"))) {
    cat(sprintf("Creating folder %s\n", here::here("processed_data")))
    dir.create(here::here("processed_data"), recursive = T)
}

if(!dir.exists(here::here("out/post_compass_plots/magnitude_plots"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/post_compass_plots/magnitude_plots")))
    dir.create(here::here("out/post_compass_plots/magnitude_plots"), recursive = T)
}

if(!dir.exists(here::here("out/post_compass_plots/fs_plots"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/post_compass_plots/fs_plots")))
    dir.create(here::here("out/post_compass_plots/fs_plots"), recursive = T)
  }
```

# Stacked heat maps by infection status

Break out the COMPASS heat maps by infection status but keep the stims stacked.

First, set cytokine order and colors for labels.

```{r}
cytokine_order_for_annotation = c("CD154", "IL2", "TNFa", "CD107a", "IL4_5_13", "IL17a", "IFNg")
cytokine_row_name_text = c("CD154", "IL-2", paste0("TNF", "\U03B1"), "CD107a", "IL-4/5/13", "IL-17a", paste0("IFN", "\U03B3")) # make sure this is in the same order as "cytokine_order_for_annotation"
group_colors <- c("Naive PRE" = "#54A8E1", "Naive POST" = "#196293", "Conv PRE" = "#EB4D4F", "Conv POST" = "#9D1213")
stim_colors <- c("S1" = "#8dd3c7", "S2" = "#ffffb3", "NCAP" = "#bebada") 
row_ann_colors <- list(`Stim` = stim_colors, `Group` = group_colors)

mergeMatricesForPlotCOMPASSResultStack_tmp <- utils::getFromNamespace("mergeMatricesForPlotCOMPASSResultStack", "COMPASS")
```

## CD4

```{r, fig.width=6, fig.height=8}
# CD4 runs
cd4_data2plot <- mergeMatricesForPlotCOMPASSResultStack_tmp(crList[CD4RunNames],
                                   threshold=0.01,
                                   minimum_dof=1,
                                   maximum_dof=Inf,
                                   row_annotation = c("Stim", "Group"),
                                   variable = "Stim")
if(save_output) {
  saveRDS(cd4_data2plot, here::here("processed_data/Merged_CD4_COMPASS_Data.rds"))
}
cd4_merged_COMPASSResult <- crList$`CD4_S1`
cd4_merged_COMPASSResult$fit$mean_gamma <- cd4_data2plot$MMerged
cd4_merged_COMPASSResult$fit$categories <- cd4_data2plot$catsMerged %>% mutate_all(~ as.numeric(as.character(.))) %>% mutate(Counts = rowSums(.))
rownames(cd4_merged_COMPASSResult$fit$categories) <- rownames(cd4_data2plot$catsMerged)
cd4_merged_COMPASSResult$data$meta <- cd4_data2plot$rowannMerged
cd4_merged_COMPASSResult$data$meta$Run_SAMPLE_ID <- rownames(cd4_merged_COMPASSResult$data$meta)
cd4_merged_COMPASSResult$data$meta$Stim <- factor(dplyr::recode(sub("^4_", "", cd4_merged_COMPASSResult$data$meta$Stim), "CD4_S1" = "S1", "CD4_S2" = "S2", "CD4_NCAP" = "NCAP"), levels = c("S1", "S2", "NCAP"))
cd4_merged_COMPASSResult$data$individual_id <- "Run_SAMPLE_ID"

# CD4 naive
# The argument "subset_keep" specifies which samples to keep based on Run_SAMPLE_ID
# Here, subset_keep = "H-" keeps all Run_SAMPLE_IDs containing "H-"
cd4_naive_heatmap <- plot.COMPASSResult.ComplexHeatmap(cr = cd4_merged_COMPASSResult,
                                                       subset_keep = "H-",
                                  row_annotation = c("Stim", "Group"),
                                  cytokine_order_for_annotation = cytokine_order_for_annotation,
                                  cytokine_row_name_text <- cytokine_row_name_text,
                                  dichotomize_by_cytokine = NULL,
                                  dichotomize_by_cytokine_color = "#999999",
                                  row_annotation_colors = row_ann_colors,
                                  staircase_cytokine_annotation = TRUE,
                                  row_gap = unit(0.03, "in"))

draw(cd4_naive_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom")

# CD4 convalescent
cd4_conv_heatmap <- plot.COMPASSResult.ComplexHeatmap(cr = cd4_merged_COMPASSResult,
                                                       subset_keep = "C-",
                                  row_annotation = c("Stim", "Group"),
                                  cytokine_order_for_annotation = cytokine_order_for_annotation,
                                  cytokine_row_name_text <- cytokine_row_name_text,
                                  dichotomize_by_cytokine = NULL,
                                  dichotomize_by_cytokine_color = "#999999",
                                  row_annotation_colors = row_ann_colors,
                                  staircase_cytokine_annotation = TRUE,
                                  row_gap = unit(0.03, "in"))

draw(cd4_conv_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom")
```

## CD8

```{r, fig.width=6, fig.height=8}
cd8_data2plot <- mergeMatricesForPlotCOMPASSResultStack_tmp(crList[CD8RunNames],
                                   threshold=0.01,
                                   minimum_dof=1,
                                   maximum_dof=Inf,
                                   row_annotation = c("Stim", "Group"),
                                   variable = "Stim")
if(save_output) {
  saveRDS(cd8_data2plot, here::here("processed_data/Merged_CD8_COMPASS_Data.rds"))
}
cd8_merged_COMPASSResult <- crList$`CD8_S1`
cd8_merged_COMPASSResult$fit$mean_gamma <- cd8_data2plot$MMerged
cd8_merged_COMPASSResult$fit$categories <- cd8_data2plot$catsMerged %>% mutate_all(~ as.numeric(as.character(.))) %>% mutate(Counts = rowSums(.))
rownames(cd8_merged_COMPASSResult$fit$categories) <- rownames(cd8_data2plot$catsMerged)
cd8_merged_COMPASSResult$data$meta <- cd8_data2plot$rowannMerged
cd8_merged_COMPASSResult$data$meta$Run_SAMPLE_ID <- rownames(cd8_merged_COMPASSResult$data$meta)
cd8_merged_COMPASSResult$data$meta$Stim <- factor(dplyr::recode(sub("^8_", "", cd8_merged_COMPASSResult$data$meta$Stim), "CD8_S1" = "S1", "CD8_S2" = "S2", "CD8_NCAP" = "NCAP"), levels = c("S1", "S2", "NCAP"))
cd8_merged_COMPASSResult$data$individual_id <- "Run_SAMPLE_ID"

# CD8 naive
cd8_naive_heatmap <- plot.COMPASSResult.ComplexHeatmap(cr = cd8_merged_COMPASSResult,
                                                       subset_keep = "H-",
                                  row_annotation = c("Stim", "Group"),
                                  cytokine_order_for_annotation = cytokine_order_for_annotation,
                                  cytokine_row_name_text <- cytokine_row_name_text,
                                  dichotomize_by_cytokine = NULL,
                                  dichotomize_by_cytokine_color = "#999999",
                                  row_annotation_colors = row_ann_colors,
                                  staircase_cytokine_annotation = TRUE,
                                  row_gap = unit(0.03, "in"))

draw(cd8_naive_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom")

# CD8 convalescent
cd8_conv_heatmap <- plot.COMPASSResult.ComplexHeatmap(cr = cd8_merged_COMPASSResult,
                                                       subset_keep = "C-",
                                  row_annotation = c("Stim", "Group"),
                                  cytokine_order_for_annotation = cytokine_order_for_annotation,
                                  cytokine_row_name_text <- cytokine_row_name_text,
                                  dichotomize_by_cytokine = NULL,
                                  dichotomize_by_cytokine_color = "#999999",
                                  row_annotation_colors = row_ann_colors,
                                  staircase_cytokine_annotation = TRUE,
                                  row_gap = unit(0.03, "in"))

draw(cd8_conv_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom")
```

# Boxplots by infection status

Note: all p-values are unadjusted.

```{r}
# Set cytokine order and colors for labels
cytokine_order_for_annotation = c("CD154", "IL2", "TNFa", "CD107a", "IL4_5_13", "IL17a", "IFNg")
cytokine_row_name_text = c("CD154", "IL-2", paste0("TNF", "\U03B1"), "CD107a", "IL-4/5/13", "IL-17a", paste0("IFN", "\U03B3")) # make sure this is in the same order as "cytokine_order_for_annotation"
naive_group_colors <- c("Naive PRE" = "#54A8E1", "Naive POST" = "#196293")
conv_group_colors <- c("Conv PRE" = "#EB4D4F", "Conv POST" = "#9D1213")
```

## CD4 

```{r}
# CD4 naive
cd4_naive_bg_corr_boxplots <- pmap(.l = list(CD4RunNames, 
                                   c("CD4+", "CD4+", "CD4+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.005), c(-0.0015, 0.004)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-", n, parentSubset = p, add_legend = T, paired = TRUE, 
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive PRE", "Naive POST"),
                                                        group_by_colors=naive_group_colors,
                                                        group_by_labels = c("Naive PRE", "Naive POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd4_naive_bg_corr_boxplots) <- CD4RunNames

if(save_output) {
  saveRDS(cd4_naive_bg_corr_boxplots, here::here("processed_data/CD4_N_make_boxplot_for_COMPASS_run_list.rds"))
}

# CD4 convalescent
cd4_conv_bg_corr_boxplots <- pmap(.l = list(CD4RunNames, 
                                   c("CD4+", "CD4+", "CD4+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.005), c(-0.0015, 0.004)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "C-", n, parentSubset = p, add_legend = T, paired = TRUE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Conv PRE", "Conv POST"),
                                                        group_by_colors=conv_group_colors,
                                                        group_by_labels = c("Conv PRE", "Conv POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd4_conv_bg_corr_boxplots) <- CD4RunNames

if(save_output) {
  saveRDS(cd4_conv_bg_corr_boxplots, here::here("processed_data/CD4_C_make_boxplot_for_COMPASS_run_list.rds"))
}
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_naive_bg_corr_boxplots[["CD4_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_conv_bg_corr_boxplots[["CD4_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_naive_bg_corr_boxplots[["CD4_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_conv_bg_corr_boxplots[["CD4_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_naive_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_conv_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
```

## CD8 

```{r}
# CD8 naive
cd8_naive_bg_corr_boxplots <- pmap(.l = list(CD8RunNames,
                                   c("CD8+", "CD8+", "CD8+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.0025), c(-0.0015, 0.003)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-", n, parentSubset = p, add_legend = T, paired = TRUE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive PRE", "Naive POST"),
                                                        group_by_colors=naive_group_colors,
                                                        group_by_labels = c("Naive PRE", "Naive POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd8_naive_bg_corr_boxplots) <- CD8RunNames

# CD8 convalescent
cd8_conv_bg_corr_boxplots <- pmap(.l = list(CD8RunNames,
                                   c("CD8+", "CD8+", "CD8+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.0025), c(-0.0015, 0.003)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "C-", n, parentSubset = p, add_legend = T, paired = TRUE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Conv PRE", "Conv POST"),
                                                        group_by_colors=conv_group_colors,
                                                        group_by_labels = c("Conv PRE", "Conv POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd8_conv_bg_corr_boxplots) <- CD8RunNames
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_naive_bg_corr_boxplots[["CD8_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_conv_bg_corr_boxplots[["CD8_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_naive_bg_corr_boxplots[["CD8_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_conv_bg_corr_boxplots[["CD8_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_naive_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_conv_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
```

# Boxplots comparing PRE

Note: all p-values are unadjusted.

```{r}
# Set cytokine order and colors for labels
cytokine_order_for_annotation = c("CD154", "IL2", "TNFa", "CD107a", "IL4_5_13", "IL17a", "IFNg")
cytokine_row_name_text = c("CD154", "IL-2", paste0("TNF", "\U03B1"), "CD107a", "IL-4/5/13", "IL-17a", paste0("IFN", "\U03B3")) # make sure this is in the same order as "cytokine_order_for_annotation"
pre_group_colors <- c("Naive PRE" = "#54A8E1", "Conv PRE" = "#EB4D4F")
```

## CD4

```{r}
# CD4 naive and convalescent pre
cd4_pre_bg_corr_boxplots <- pmap(.l = list(CD4RunNames, 
                                   c("CD4+", "CD4+", "CD4+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.0035), c(-0.0015, 0.004)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-1|C-1", n, parentSubset = p, add_legend = T, paired = FALSE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive PRE", "Conv PRE"),
                                                        group_by_colors=pre_group_colors,
                                                        group_by_labels = c("Naive PRE", "Conv PRE"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd4_pre_bg_corr_boxplots) <- CD4RunNames
if(save_output) {
  saveRDS(cd4_pre_bg_corr_boxplots, here::here("processed_data/CD4_pre_make_boxplot_for_COMPASS_run_list.rds"))
}
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_pre_bg_corr_boxplots[["CD4_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_pre_bg_corr_boxplots[["CD4_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_pre_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
```

## CD8 

```{r}
# CD8 naive and convalescent pre
cd8_pre_bg_corr_boxplots <- pmap(.l = list(CD8RunNames,
                                   c("CD8+", "CD8+", "CD8+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.0025), c(-0.0015, 0.003)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-1|C-1", n, parentSubset = p, add_legend = T, paired = FALSE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive PRE", "Conv PRE"),
                                                        group_by_colors=pre_group_colors,
                                                        group_by_labels = c("Naive PRE", "Conv PRE"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd8_pre_bg_corr_boxplots) <- CD8RunNames
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_pre_bg_corr_boxplots[["CD8_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_pre_bg_corr_boxplots[["CD8_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_pre_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
```

# Boxplots comparing POST

Note: all p-values are unadjusted.

```{r}
# Set cytokine order and colors for labels
cytokine_order_for_annotation = c("CD154", "IL2", "TNFa", "CD107a", "IL4_5_13", "IL17a", "IFNg")
cytokine_row_name_text = c("CD154", "IL-2", paste0("TNF", "\U03B1"), "CD107a", "IL-4/5/13", "IL-17a", paste0("IFN", "\U03B3")) # make sure this is in the same order as "cytokine_order_for_annotation"
post_group_colors <- c("Naive POST" = "#196293", "Conv POST" = "#9D1213")
```

## CD4

```{r}
# CD4 naive and convalescent post
cd4_post_bg_corr_boxplots <- pmap(.l = list(CD4RunNames, 
                                   c("CD4+", "CD4+", "CD4+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.005), c(-0.0015, 0.004)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-2|C-2", n, parentSubset = p, add_legend = T, paired = FALSE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive POST", "Conv POST"),
                                                        group_by_colors=post_group_colors,
                                                        group_by_labels = c("Naive POST", "Conv POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd4_post_bg_corr_boxplots) <- CD4RunNames
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_post_bg_corr_boxplots[["CD4_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_post_bg_corr_boxplots[["CD4_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd4_post_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
```

## CD8 

```{r}
# CD8 naive and convalescent post
cd8_post_bg_corr_boxplots <- pmap(.l = list(CD8RunNames,
                                   c("CD8+", "CD8+", "CD8+"),
                                   list(c(-0.0015, 0.003), c(-0.0015, 0.0025), c(-0.0015, 0.003)),
                                   c(7, 7, 7),
                                   c(1.3, 1.3, 1.3),
                                   list(NULL, NULL, NULL),
                                   c("S1", "S2", "NCAP")),
                         .f = function(n, p, y, p_text_size, left_cats_heatmap_padding, dichotomize_by_cytokine, main_title) {
                           # "staircase effect" for categories df heatmap gets done automatically, unlike in heatmap
                           # Returned plot is gtable
                           make_boxplot_for_COMPASS_run(crList[[n]], subset_keep = "H-2|C-2", n, parentSubset = p, add_legend = T, paired = FALSE,
                                                        current_ylim = y, include_0_line=T, 
                                                        cytokine_order_for_annotation=cytokine_order_for_annotation,
                                                        cytokine_row_name_text = cytokine_row_name_text, 
                                                        p_text_size=p_text_size,
                                                        group_by_order=c("Naive POST", "Conv POST"),
                                                        group_by_colors=post_group_colors,
                                                        group_by_labels = c("Naive POST", "Conv POST"),
                                                        group_by_colname="Group",
                                                        zeroed_BgCorr_stats = FALSE, zeroed_BgCorr_plot = FALSE,
                                                        dichotomize_by_cytokine = dichotomize_by_cytokine, dichotomize_by_cytokine_color="#999999",
                                                        cats_heatmap_left_padding=left_cats_heatmap_padding, main_title = main_title,
                                                        save_test_results=TRUE, output_folder=here::here("processed_data"), adj_pval = FALSE)
                         })
names(cd8_post_bg_corr_boxplots) <- CD8RunNames
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_post_bg_corr_boxplots[["CD8_S1"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_post_bg_corr_boxplots[["CD8_S2"]]$Plot)
```

```{r, fig.width=10, fig.height=7}
grid.draw(cd8_post_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
```

# FS and PFS

```{r}
fs_pfs_df <- lapply(c(CD4RunNames, CD8RunNames), function(n) {
  as.data.frame(COMPASS::FunctionalityScore(crList[[n]])) %>%
    rownames_to_column("SAMPLE ID") %>% 
    rename(FS = `COMPASS::FunctionalityScore(crList[[n]])`) %>% 
    left_join(as.data.frame(COMPASS::PolyfunctionalityScore(crList[[n]])) %>%
      rownames_to_column("SAMPLE ID") %>% 
      rename(PFS = `COMPASS::PolyfunctionalityScore(crList[[n]])`),
      by = "SAMPLE ID") %>% 
    mutate(COMPASS_run = n)
  } ) %>% 
  data.table::rbindlist() %>% 
  separate(COMPASS_run, into = c("parent", "Stim"), remove = F, extra = "merge") %>% 
  left_join(crList$CD4_S1$data$meta %>%
              dplyr::select(`SAMPLE ID`, Group, Timepoint,
                            `PATIENT ID`))

i1 <- grepl("H", fs_pfs_df$`PATIENT ID`) 
i2 <- grepl("C", fs_pfs_df$`PATIENT ID`) 
fs_pfs_df$Infection_Status[i1] <- "Naive"
fs_pfs_df$Infection_Status[i2] <- "Conv"

fs_pfs_df$Infection_Status <- factor(fs_pfs_df$Infection_Status, levels = c("Naive", "Conv"))
fs_pfs_df$Group <- factor(fs_pfs_df$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

Write out the functionality scores to a file

```{r}
if(save_output) {
  fs_pfs_dat_2save <- fs_pfs_df %>% 
    # Convert the numeric columns to character type after rounding to 20 digits. This will help ensure consistency when writing and reading the data to a file
    mutate_at(vars("FS", "PFS"), ~ format(., digits = 20))
  
  write.csv(fs_pfs_dat_2save,
            here::here("processed_data/HAARVIVAC_COMPASS_FS_PFS.csv"), row.names = F)
}
```

Set colors 

```{r}
fill_colors <- list("Naive" = c("Naive PRE" = "#54A8E1", "Naive POST" = "#196293"),
                    "Conv" = c("Conv PRE" = "#EB4D4F", "Conv POST" = "#9D1213"),
                    "PRE" = c("Naive PRE" = "#54A8E1", "Conv PRE" = "#EB4D4F"),
                    "POST" = c("Naive POST" = "#196293", "Conv POST" = "#9D1213"),
                    "Naive PRE" = c("Naive PRE" = "#54A8E1"),
                    "Conv PRE" = c("Conv PRE" = "#EB4D4F"))
```

Use Wilcoxon signed-rank test to compare across stims within a donor. 
Show the medians in each group.
Note: p-values are adjusted

## FS Plots

```{r}
naive_pre_fs_plot <- fs_pfs_plot(df = fs_pfs_df, FS_or_PFS = "FS", cd4_or_cd8 = "CD4",
                                 group = "Naive PRE", fill_colors = fill_colors, group_by_colname = "Group")
naive_pre_fs_plot

conv_pre_fs_plot <- fs_pfs_plot(df = fs_pfs_df, FS_or_PFS = "FS", cd4_or_cd8 = "CD4",
                                group = "Conv PRE", fill_colors = fill_colors, group_by_colname = "Group")
conv_pre_fs_plot
```

## PFS Plots

```{r}
naive_pre_pfs_plot <- fs_pfs_plot(df = fs_pfs_df, FS_or_PFS = "PFS", cd4_or_cd8 = "CD4",
                                  group = "Naive PRE", fill_colors = fill_colors, group_by_colname = "Group")
naive_pre_pfs_plot

conv_pre_pfs_plot <- fs_pfs_plot(df = fs_pfs_df, FS_or_PFS = "PFS", cd4_or_cd8 = "CD4",
                                 group = "Conv PRE", fill_colors = fill_colors, group_by_colname = "Group")
conv_pre_pfs_plot
```

# Split FS plots

## Naive

```{r, fig.width=4, fig.height=5}
n_fs_plots <- pmap(list(rep(c("S1", "S2", "NCAP"), times = 2),
                           rep(c("CD4", "CD8"), each = 3)),
                           function(current_stim, cd4_or_cd8) {
                             split_fs_pfs_plot(df = fs_pfs_df, "FS", current_stim, cd4_or_cd8, compare_naive = TRUE, compare_intra = TRUE,
                                               fill_colors = fill_colors, group_by_colname = "Group", groups_to_compare = "Naive", ylim = c(0, 0.125))
                           })
names(n_fs_plots) <- paste0(rep(c("CD4", "CD8"), each = 3), "_", rep(c("S1", "S2", "NCAP"), times = 2))
n_fs_plots
```

## Conv

```{r, fig.width=4, fig.height=5}
c_fs_plots <- pmap(list(rep(c("S1", "S2", "NCAP"), times = 2),
                           rep(c("CD4", "CD8"), each = 3)),
                           function(current_stim, cd4_or_cd8) {
                             split_fs_pfs_plot(df = fs_pfs_df, "FS", current_stim, cd4_or_cd8, compare_conv = TRUE, compare_intra = TRUE,
                                               fill_colors = fill_colors, group_by_colname = "Group", groups_to_compare = "Conv", ylim = c(0, 0.1125))
                           })
names(c_fs_plots) <- paste0(rep(c("CD4", "CD8"), each = 3), "_", rep(c("S1", "S2", "NCAP"), times = 2))
c_fs_plots
```

## PRE

```{r, fig.width=4, fig.height=5}
pre_fs_plots <- pmap(list(rep(c("S1", "S2", "NCAP"), times = 2),
                           rep(c("CD4", "CD8"), each = 3)),
                           function(current_stim, cd4_or_cd8) {
                             split_fs_pfs_plot(df = fs_pfs_df, "FS", current_stim, cd4_or_cd8, compare_PRE = TRUE, 
                                               fill_colors = fill_colors, group_by_colname = "Group", groups_to_compare = "PRE")
                           })
names(pre_fs_plots) <- paste0(rep(c("CD4", "CD8"), each = 3), "_", rep(c("S1", "S2", "NCAP"), times = 2))
pre_fs_plots
```

## POST

```{r, fig.width=4, fig.height=5}
post_fs_plots <- pmap(list(rep(c("S1", "S2"), times = 2),
                           rep(c("CD4", "CD8"), each = 2)),
                           function(current_stim, cd4_or_cd8) {
                             split_fs_pfs_plot(df = fs_pfs_df, "FS", current_stim, cd4_or_cd8, compare_POST = TRUE, 
                                               fill_colors = fill_colors, group_by_colname = "Group", groups_to_compare = "POST", ylim = c(0, 0.125))
                           })
names(post_fs_plots) <- paste0(rep(c("CD4", "CD8"), each = 2), "_", rep(c("S1", "S2"), times = 2))
post_fs_plots
```

# Save figures

## Heat maps

```{r save heatmaps pdf}
if(save_output) {
  cairo_pdf(file=here::here("out/post_compass_plots/CD4_naive_COMPASS_Heatmap.pdf"), width=5, height=8,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(draw(cd4_naive_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom"))
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/CD4_conv_COMPASS_Heatmap.pdf"), width=5, height=8,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica
  print(draw(cd4_conv_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom"))
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/CD8_naive_COMPASS_Heatmap.pdf"), width=5, height=8,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica
  print(draw(cd8_naive_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom"))
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/CD8_conv_COMPASS_Heatmap.pdf"), width=5, height=8,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica
  print(draw(cd8_conv_heatmap, gap = unit(0.1, "in"), merge_legends = T, heatmap_legend_side = "bottom", annotation_legend_side = "bottom"))
  dev.off()
}
```

## Boxplots

```{r save boxplots pdf}
if(save_output) {
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD4_naive_boxplots.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd4_naive_bg_corr_boxplots[["CD4_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd4_naive_bg_corr_boxplots[["CD4_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd4_naive_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD4_conv_boxplots.pdf"), width=8, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd4_conv_bg_corr_boxplots[["CD4_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd4_conv_bg_corr_boxplots[["CD4_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd4_conv_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD8_naive_boxplots.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd8_naive_bg_corr_boxplots[["CD8_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd8_naive_bg_corr_boxplots[["CD8_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd8_naive_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD8_conv_boxplots.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd8_conv_bg_corr_boxplots[["CD8_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd8_conv_bg_corr_boxplots[["CD8_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd8_conv_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
  dev.off()

  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD4_pre_boxplots.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd4_pre_bg_corr_boxplots[["CD4_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd4_pre_bg_corr_boxplots[["CD4_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd4_pre_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD8_pre_boxplots.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd8_pre_bg_corr_boxplots[["CD8_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd8_pre_bg_corr_boxplots[["CD8_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd8_pre_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD4_post_boxplot.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd4_post_bg_corr_boxplots[["CD4_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd4_post_bg_corr_boxplots[["CD4_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd4_post_bg_corr_boxplots[["CD4_NCAP"]]$Plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/magnitude_plots/CD8_post_boxplot.pdf"), width=10, height=6,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  grid.draw(cd8_post_bg_corr_boxplots[["CD8_S1"]]$Plot)
  grid.newpage()
  grid.draw(cd8_post_bg_corr_boxplots[["CD8_S2"]]$Plot)
  grid.newpage()
  grid.draw(cd8_post_bg_corr_boxplots[["CD8_NCAP"]]$Plot)
  dev.off()
}
```

## FS plots

```{r save split FS plots pdf}
if(save_output) {
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/naive_pre_fs_plots.pdf"), width=3.5, height=4,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(naive_pre_fs_plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/conv_pre_fs_plots.pdf"), width=3.5, height=4,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(conv_pre_fs_plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/naive_fs_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(n_fs_plots)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/conv_fs_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(c_fs_plots)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/pre_fs_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(pre_fs_plots)
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/fs_plots/post_fs_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(post_fs_plots)
  dev.off()
}
```
