==> scripts/05_UMAP_on_CD4_COMPASS_subsets.Rmd <==
---
title: "HAARVIVAC ICS Dimensionality Reduction on CD4+ COMPASS subsets"
author: "Jolie Phan"
date: "version `r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Run UMAP on CD4+ events which express a COMPASS subset. 
What are the memory and activation profiles of antigen-specific T cells?

# Load libraries

```{r, message=FALSE}
library(openCyto)
library(CytoML) 
library(flowCore) 
library(flowWorkspace) 
library(here)
library(tidyverse)
library(uwot)
library(ggplot2)
library(scales)
library(cowplot)
library(hues)
library(ggrepel)
library(ggpubr)
library(ggrastr)
source(here::here("scripts/Helper_Functions.R"))
```

# Load data

```{r}
date <- 20220302
save_output <- FALSE
rerun_dimred <- FALSE
save_fig <- TRUE
gs <- load_gs(here::here("out/GatingSets/HAARVIVAC_GatingSet_with_COMPASS_Subsets"))

if(!dir.exists(here::here("out/UMAP"))) {
  cat(sprintf("Creating folder %s\n", here::here("out/UMAP")))
  dir.create(here::here("out/UMAP"), recursive = T)
}
```

# Extract mfi data

```{r}
cd4_gates_for_dimred <- c(
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD107a+", "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD154+", 
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+", "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL2+", 
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL17a+", "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL4_5_13+", 
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/TNFa+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CCR7+", "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD45RA+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD38+", "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/HLADR+")

cd4_cytokine_gates <- c(
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD107a+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD154+", 
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL2+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL17a+",
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IL4_5_13+", 
  "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/TNFa+"
)
```

```{r}
# Get counts for CD4_COMPASS_Subsets
cd4_compass_subsets_parentGate <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets"

pop_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_subsets_parentGate), by = c("rowname" = "name")) %>%
  dplyr::rename(CD4_COMPASS_Subsets = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", Stim, Group, CD4_COMPASS_Subsets) %>%
  dplyr::filter(Stim != "SEB")
```

```{r}
# Extract data for dimensionality reduction (not actually sampling)
call_sampleGatingHierarchy_for_cd4 <- function(currentSampleName) {
  sampleGatingHierarchy(gs[[currentSampleName]], cd4_compass_subsets_parentGate, n = NULL, otherGates = cd4_gates_for_dimred)
}

cd4_compass_subsets_data <- map_dfr(pop_counts$rowname, call_sampleGatingHierarchy_for_cd4) 
dim(cd4_compass_subsets_data)
knitr::kable(head(cd4_compass_subsets_data))
```

# Perform Dimensionality Reduction

```{r}
cols_4_dimred <- c("CD3", "CD8b", "CD4", "TNFa", "CD107a", "CD154", "IL2", "IL17a", "IL4_5_13", "IFNg", "CCR7", "CD45RA", "CD38", "HLADR")
# Scale CD4 COMPASS subset counts 
cd4.scaled_dimred_input <- cd4_compass_subsets_data %>%
  dplyr::select("EXPERIMENT NAME", all_of(cols_4_dimred)) %>%
  group_by(`EXPERIMENT NAME`) %>%
  nest() %>%
  ungroup() %>%
  mutate(data = lapply(data, function(df) {as.data.frame(scale(as.matrix(df)))})) %>%
  unnest(cols = c(data)) %>%
  rename_at(vars(all_of(cols_4_dimred)),function(x) paste0(x,".scaled")) %>% 
  dplyr::select(-"EXPERIMENT NAME")

cd4_compass_subsets_data <- cbind(cd4_compass_subsets_data, cd4.scaled_dimred_input)

if(rerun_dimred) {
  print("Running UMAP")
  set.seed(date)
  print(Sys.time())
  cd4_compass_subsets_dimred_out <- cd4_compass_subsets_data %>%
    # Run CD3, co-receptor, cytokine, memory, and activation markers through UMAP
    dplyr::select(all_of(paste0(cols_4_dimred, ".scaled"))) %>%
    uwot::umap(spread = 9, min_dist = 0.02, n_threads = 7)
  print(Sys.time())
  cd4_compass_subsets_w_umap <- cbind(as.data.frame(cd4_compass_subsets_dimred_out) %>%
    dplyr::rename(x.umap = V1, y.umap = V2),
  cd4_compass_subsets_data)
  if(save_output) {
    saveRDS(cd4_compass_subsets_w_umap, here::here(sprintf("out/UMAP/%s_ICS_CD4_COMPASS_Subsets_UMAP_Unsampled.rds", date)))
    }
} else {
  # Assuming UMAP results are already saved
  print("Loading saved UMAP run")
  cd4_compass_subsets_w_umap <- readRDS(here::here(sprintf("out/UMAP/%s_ICS_CD4_COMPASS_Subsets_UMAP_Unsampled.rds", date)))
}
```

# Plot UMAP results

```{r}
set.seed(date)
cd4_compass_subsets_w_umap <- cd4_compass_subsets_w_umap[sample(nrow(cd4_compass_subsets_w_umap), nrow(cd4_compass_subsets_w_umap)),]
```

```{r}
# Arial font setup. Downloaded afms from https://github.com/microsoft/microsoft-r-open/tree/ec3fd89e5fb5794bd8149905c134ad801bb61800
Arial <- Type1Font(family = "Arial",
                   metrics = c(here::here("data/Arial_afm/ArialMT.afm"),
                               here::here("data/Arial_afm/ArialMT-Bold.afm"), 
                              here::here("data/Arial_afm/ArialMT-Italic.afm"),                          here::here("data/Arial_afm/ArialMT-BoldItalic.afm")))
windowsFonts("Arial" = windowsFont("Arial"))
pdfFonts(Arial = Arial)

boolColorScheme <- c("FALSE" = "#9E9E9E", "TRUE" = "#C7254A")

cd4_compass_subsets_w_umap <- cd4_compass_subsets_w_umap %>%
  mutate(cytokine_degree = rowSums(dplyr::select(., all_of(cd4_cytokine_gates))))

stim_labs <- c("DMSO", "S1", "S2", "NCAP")
names(stim_labs) <- c("DMSO", "S1", "S2", "NCAP")
group_labs <- c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST")
names(group_labs) <- c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST")

base_dimred_plot <- function(currentColumn, pointSize = 0.02, colorScheme = NA) {
  p <- ggplot(cd4_compass_subsets_w_umap, aes(x=x.umap, y=y.umap,
                                       colour=if(currentColumn %in% c("EXPERIMENT NAME", "SAMPLE ID")) {
                                         factor(!!as.name(currentColumn))
                                         } else {
                                           as.logical(!!as.name(currentColumn))
                                           })) +
    geom_point_rast(shape=20, alpha=0.8, size=pointSize) +
    facet_grid(Group ~ Stim, switch="y",
               labeller = labeller(Group = group_labs, Stim = stim_labs)) +
    theme_bw() +
    theme(plot.title=element_text(hjust = 0.5, size=22, face="bold"),
          axis.text=element_blank(),
          axis.line=element_blank(),
          axis.ticks=element_blank(),
          axis.title=element_blank(),
          strip.text=element_text(face="bold", size=10),
          panel.grid.major = element_blank(),
          legend.title=element_text(face="bold", size=14),
          strip.text.x = element_text(margin = margin(0.15,0,0.15,0, "cm")))
  if(!anyNA(colorScheme)) {
    p <- p + scale_color_manual(values = colorScheme)
  }
  if(currentColumn %in% c("EXPERIMENT NAME", "SAMPLE ID")) {
    p <- p + labs(color = currentColumn) +
      guides(colour = guide_legend(override.aes = list(size=7))) +
      theme(legend.title = element_text(size=10),
            legend.text = element_text(size=10),
            legend.position = "bottom") +
      scale_colour_manual(values=as.character(iwanthue(length(unique(cd4_compass_subsets_w_umap[,currentColumn])))))
  } else {
    p <- p + theme(legend.position = "none")
  }
  p
}
```

## Look for Batch and Patient sub-clustering

```{r, fig.width=10, fig.height=8}
base_dimred_plot("EXPERIMENT NAME")
```

```{r, fig.width=10, fig.height=8}
base_dimred_plot("SAMPLE ID")
```

## Individual marker expression localization

```{r, fig.width=10, fig.height=8, eval=FALSE, include=FALSE}
for(cg in cd4_gates_for_dimred) {
  print(base_dimred_plot(cg, colorScheme = boolColorScheme) +
    labs(title = sprintf("CD4+ COMPASS Subset+ UMAP\nColored by %s", sub(".*\\/([^(\\/)]+)", "\\1", cg))))
}
```

## Color by degree

```{r, fig.width=10, fig.height=8, eval=FALSE, include=FALSE}
table(cd4_compass_subsets_w_umap$cytokine_degree)
# myPalette <- colorRampPalette(GetColors(20, scheme = "smooth rainbow", start = 0.25, end = 0.95))
myPalette <- viridis_pal()
sc <- scale_colour_gradientn(colours = myPalette(100), limits=c(1, 5))
currentColumn <- "cytokine_degree"
pointSize <- 0.02
ggplot(cd4_compass_subsets_w_umap, aes(x=x.umap, y=y.umap, colour=!!as.name(currentColumn))) +
  geom_point_rast(shape=20, alpha=0.8, size=pointSize) +
  facet_grid(Group ~ Stim, switch="y",
             labeller = labeller(Group = group_labs, Stim = stim_labs)) +
  labs(colour="Cytokine\nDegree",
       title="CD4+ COMPASS Subset+ UMAP\nColored by Cytokine Polyfunctionality") +
  theme_bw() +
  theme(plot.title=element_text(hjust = 0.5, size=22, face="bold"),
        axis.text=element_blank(),
        axis.line=element_blank(),
        axis.ticks=element_blank(),
        axis.title=element_blank(),
        strip.text=element_text(face="bold", size=10),
        panel.grid.major = element_blank(),
        legend.title=element_text(face="bold", size=15),
        legend.text = element_text(size=13),
        strip.text.x = element_text(margin = margin(0.15,0,0.15,0, "cm")),
        legend.position = "bottom") +
  sc
```

# Plot activation, memory, and PolyF results

## Add memory and activation categories

```{r}
# Add memory column
getMemoryCategory <- function(CD45RA, CCR7) {
  if(CD45RA & CCR7) {
    "Naive"
  } else if(CD45RA & !CCR7) {
    "TEMRA"
  } else if(!CD45RA & CCR7) {
    "TCM"
  } else if(!CD45RA & !CCR7) {
    "TEM"
  } else {
    "Uncategorized"
  }
}
cd4_compass_subsets_w_umap$MemoryCategory <- pmap_chr(.l = list(cd4_compass_subsets_w_umap$`/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD45RA+`,
                                                               cd4_compass_subsets_w_umap$`/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CCR7+`),
                                                     .f = getMemoryCategory)
cd4_compass_subsets_w_umap$MemoryCategory <- factor(cd4_compass_subsets_w_umap$MemoryCategory, levels = c("Naive", "TCM", "TEMRA", "TEM"))

# Add activation column
getActivationCategory <- function(HLADR, CD38) {
  if(HLADR & CD38) {
    "HLA-DR+/CD38+"
  } else if(!HLADR & !CD38) {
    "HLA-DR-/CD38-"
  } else if(HLADR & !CD38) {
    "HLA-DR+/CD38-"
  } else if(!HLADR & CD38) {
    "HLA-DR-/CD38+"
  }
}

cd4_compass_subsets_w_umap$ActivationCategory <- pmap_chr(.l = list(cd4_compass_subsets_w_umap$`/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/HLADR+`,
                                                               cd4_compass_subsets_w_umap$`/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD38+`),
                                                     .f = getActivationCategory)
cd4_compass_subsets_w_umap$ActivationCategory <- factor(cd4_compass_subsets_w_umap$ActivationCategory, levels = c("HLA-DR+/CD38+", "HLA-DR-/CD38+", "HLA-DR+/CD38-", "HLA-DR-/CD38-"))
```

## Subset data 

```{r}
naive_data <- cd4_compass_subsets_w_umap %>%
  dplyr::filter(Group == "Naive PRE" | Group == "Naive POST")

# naive_pre_data <- cd4_compass_subsets_w_umap %>%
#   dplyr::filter(Group == "Naive PRE")
# 
# naive_post_data <- cd4_compass_subsets_w_umap %>%
#   dplyr::filter(Group == "Naive POST")

conv_data <- cd4_compass_subsets_w_umap %>%
  dplyr::filter(Group == "Conv PRE" | Group == "Conv POST")

# conv_pre_data <- cd4_compass_subsets_w_umap %>%
#   dplyr::filter(Group == "Conv PRE")
# 
# conv_post_data <- cd4_compass_subsets_w_umap %>%
#   dplyr::filter(Group == "Conv POST")
```

## Set ggplot settings

```{r}
pointSize <- 0.02

# Naive 
naive_gg_themes <- ggplot(naive_data, aes(x=x.umap, y=y.umap)) +
  scale_x_continuous(limits=range(naive_data$x.umap)*1.16) +
  scale_y_continuous(limits=range(naive_data$y.umap)*1.11) +
  theme_bw() +
  theme(plot.title=element_text(hjust = 0.5, size=21),
        axis.text=element_blank(),
        axis.line=element_blank(),
        axis.ticks=element_blank(),
        axis.title=element_blank(),
        panel.grid = element_blank(),
        legend.position = "right",
        legend.justification = "left",
        legend.title.align = 0.5,
        legend.text = element_text(size = 17, color = "black"),
        legend.title = element_text(size = 20, color = "black"),
        legend.spacing.x = unit(0.5, "cm"),
        plot.margin = margin(1, 0, 0, 0, unit = "pt")) +
  guides(colour = guide_legend(override.aes = list(size=12)))

# Conv
conv_gg_themes <- ggplot(conv_data, aes(x=x.umap, y=y.umap)) +
  scale_x_continuous(limits=range(conv_data$x.umap)*1.16) +
  scale_y_continuous(limits=range(conv_data$y.umap)*1.11) +
  theme_bw() +
  theme(plot.title=element_text(hjust = 0.5, size=21),
        axis.text=element_blank(),
        axis.line=element_blank(),
        axis.ticks=element_blank(),
        axis.title=element_blank(),
        panel.grid = element_blank(),
        legend.position = "right",
        legend.justification = "left",
        legend.title.align = 0.5,
        legend.text = element_text(size = 17, color = "black"),
        legend.title = element_text(size = 20, color = "black"),
        legend.spacing.x = unit(0.5, "cm"),
        plot.margin = margin(1, 0, 0, 0, unit = "pt")) +
  guides(colour = guide_legend(override.aes = list(size=12))) 
```

## Memory 

Colored by gate membership (TEMRA, TEM, TCM, Naive)

Set colors.

```{r}
# pal <- GetColors(20, scheme = "smooth rainbow", start = 0.25, end = 0.95)
pal <- viridis_pal()(20)

# TEM is grey, TEMRA is yellow, TCM is blue, and Naive is green
memColorScheme <- c("TEM" = "#999999", "TEMRA" = pal[20], "TCM" = pal[7], "Naive" = pal[16]) 
```

Get plots.

```{r}
naive_mem_plot_colored_by_gate <- naive_gg_themes +
  geom_point_rast(aes(colour=naive_data[,"MemoryCategory"]),
               shape=20, alpha=0.8, size=pointSize) +
  scale_color_manual(values = memColorScheme) +
  labs(colour = "Memory")

conv_mem_plot_colored_by_gate <- conv_gg_themes +
  geom_point_rast(aes(colour=conv_data[,"MemoryCategory"]),
               shape=20, alpha=0.8, size=pointSize) +
  scale_color_manual(values = memColorScheme) +
  labs(colour = "Memory")
```

## Activation 

Colored by HLADR+CD38+ and HLADR-CD38-

Set colors.

```{r}
pal <- viridis_pal()(20)

# HLADR-/CD38- is grey, HLADR+/CD38+ is green, HLADR+/CD38- is yellow, and HLADR-/CD38+ is blue
actColorScheme <- c("HLA-DR-/CD38-" = "#999999", "HLA-DR+/CD38+" = pal[16], "HLA-DR+/CD38-" = pal[20], "HLA-DR-/CD38+" = pal[7])
```

Get plots.

```{r}
naive_act_plot_colored_by_gate <- naive_gg_themes +
  geom_point_rast(aes(colour=naive_data[,"ActivationCategory"]),
               shape=20, alpha=0.8, size=pointSize) +
  scale_color_manual(values = actColorScheme) +
  labs(colour = "Activation")

conv_act_plot_colored_by_gate <- conv_gg_themes +
  geom_point_rast(aes(colour=conv_data[,"ActivationCategory"]),
               shape=20, alpha=0.8, size=pointSize) +
  scale_color_manual(values = actColorScheme) +
  labs(colour = "Activation")
```

## PolyF

Get plots.

```{r}
myPalette <- viridis_pal()
sc_polyf <- scale_colour_gradientn(colours = myPalette(100), limits=c(1, 5))
naive_polyf_plot <- naive_gg_themes +
  geom_point_rast(aes(colour=naive_data[,"cytokine_degree"]), shape=20, alpha=0.8, size=pointSize) +
  sc_polyf +
  labs(colour = "PolyF") +
  guides(colour=guide_colorbar(title.position = "left")) +
  theme(legend.title = element_text(angle=90))

conv_polyf_plot <- conv_gg_themes +
  geom_point_rast(aes(colour=conv_data[,"cytokine_degree"]), shape=20, alpha=0.8, size=pointSize) +
  sc_polyf +
  labs(colour = "PolyF") +
  guides(colour=guide_colorbar(title.position = "left")) +
  theme(legend.title = element_text(angle=90))
```

## Naive plots

```{r}
naive_umap_list <- align_plots(naive_act_plot_colored_by_gate, naive_mem_plot_colored_by_gate, naive_polyf_plot, align = "hv")
names(naive_umap_list) <- c("act", "mem", "polyf")
```

```{r, fig.width=5, fig.height=7}
ggdraw(naive_umap_list[["polyf"]])
```

```{r, fig.width=5, fig.height=7}
ggdraw(naive_umap_list[["mem"]])
```

```{r, fig.width=5, fig.height=7}
ggdraw(naive_umap_list[["act"]])
```

## Conv plots

```{r}
conv_umap_list <- align_plots(conv_act_plot_colored_by_gate, conv_mem_plot_colored_by_gate, conv_polyf_plot, align = "hv")
names(conv_umap_list) <- c("act", "mem", "polyf")
```

```{r, fig.width=7, fig.height=5}
ggdraw(conv_umap_list[["polyf"]])
```

```{r, fig.width=7, fig.height=5}
ggdraw(conv_umap_list[["mem"]])
```

```{r, fig.width=7, fig.height=5}
ggdraw(conv_umap_list[["act"]])
```

# Save figures

```{r}
if(save_fig) {
  cairo_pdf(file=here::here("out/UMAP/CD4_naive_umap_plots.pdf"), width=7, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(ggdraw(naive_umap_list[["act"]]))
  print(ggdraw(naive_umap_list[["mem"]]))
  print(ggdraw(naive_umap_list[["polyf"]]))
  dev.off()

  cairo_pdf(file=here::here("out/UMAP/CD4_conv_umap_plots.pdf"), width=7, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(ggdraw(conv_umap_list[["act"]]))
  print(ggdraw(conv_umap_list[["mem"]]))
  print(ggdraw(conv_umap_list[["polyf"]]))
  dev.off()
}
```
