==> scripts/06_COMPASS_Subsets_Memory_and_Activation_Boxplots.Rmd <==
---
title: "COMPASS Subset CD4+ T cell Memory and Activation Boxplots"
author: "Jolie Phan"
date: "version `r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
editor_options:
  chunk_output_type: console
---

Plots frequencies of CD4+ T cells which express a COMPASS subset based on activation phenotype (HLA-DR+CD38+ and CD38+, HLADR agnostic) and memory phenotype (TCM and TEM).

Stratify by stim (S1, S2, NCAP, and DMSO).

Note: p-values are adjusted.

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.width = 4, fig.height = 5)
```

# Load libraries

```{r, message = FALSE}
library(here)
library(flowWorkspace)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(ggbeeswarm)
library(cowplot)
library(ggh4x)
source(here::here("scripts/Helper_Functions.R")) # for make_mag_plots()
```

# Load data

```{r}
save_output <- TRUE

gs <- load_gs(here::here("out/GatingSets/HAARVIVAC_GatingSet_with_COMPASS_Subsets"))

if(!dir.exists(here::here("out/post_compass_plots/activation_memory_plots"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/post_compass_plots/activation_memory_plots")))
    dir.create(here::here("out/post_compass_plots/activation_memory_plots"), recursive = T)
}
```

# Set font

```{r}
# Arial font setup. Downloaded afms from https://github.com/microsoft/microsoft-r-open/tree/ec3fd89e5fb5794bd8149905c134ad801bb61800
Arial <- Type1Font(family = "Arial",
                   metrics = c(here::here("data/Arial_afm/ArialMT.afm"),
                               here::here("data/Arial_afm/ArialMT-Bold.afm"),
                               here::here("data/Arial_afm/ArialMT-Italic.afm"),
                               here::here("data/Arial_afm/ArialMT-BoldItalic.afm")))
windowsFonts("Arial" = windowsFont("Arial"))
pdfFonts(Arial = Arial)
```

# Extract frequencies

Exclude SEB.

```{r}
cd4_compass_act_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets/HLADR+CD38+"

cd4_compass_cd38_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets/CD38+"

cd4_compass_hladr_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets/HLADR+"

cd4_compass_cm_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets/TCM"

cd4_compass_em_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD4_COMPASS_Subsets/TEM"
```

## HLA-DR+CD38+

```{r}
act_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_act_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

act_counts$Group <- factor(act_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## CD38+

```{r}
cd38_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_cd38_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

cd38_counts$Group <- factor(cd38_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## HLA-DR+

```{r}
hladr_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_hladr_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

hladr_counts$Group <- factor(hladr_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## Central memory

```{r}
cm_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_cm_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

cm_counts$Group <- factor(cm_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## Effector memory

```{r}
em_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_compass_em_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

em_counts$Group <- factor(em_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

# Plots

```{r}
stims <- c("S1", "S2")

fill_colors <- list("Naive" = c("Naive PRE" = "#54A8E1", "Naive POST" = "#196293"),
                    "Conv" = c("Conv PRE" = "#EB4D4F", "Conv POST" = "#9D1213"),
                    "PRE" = c("Naive PRE" = "#54A8E1", "Conv PRE" = "#EB4D4F"),
                    "POST" = c("Naive POST" = "#196293", "Conv POST" = "#9D1213"))

y_axis_size <- 15
```

## HLA-DR+CD38+

### Naive PRE vs Naive POST

```{r}
naive_act_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(act_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(naive_act_plots) <- stims
naive_act_plots
```

### Conv PRE vs Conv POST

```{r}
conv_act_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(act_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"), 
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(conv_act_plots) <- stims
conv_act_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_act_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(act_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", groups_to_compare = c("Naive PRE", "Conv PRE"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(pre_act_plots) <- stims
pre_act_plots
```

### Naive POST vs Conv POST

```{r}
post_act_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(act_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(post_act_plots) <- stims
post_act_plots
```

## CD38+

### Naive PRE vs Naive POST

```{r}
naive_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(naive_cd38_plots) <- stims
naive_cd38_plots
```

### Conv PRE vs Conv POST

```{r}
conv_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15,  ylim = c(0, 65))
                               })
names(conv_cd38_plots) <- stims
conv_cd38_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", groups_to_compare = c("Naive PRE", "Conv PRE"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(pre_cd38_plots) <- stims
pre_cd38_plots
```

### Naive POST vs Conv POST

```{r}
post_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "CD38+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(post_cd38_plots) <- stims
post_cd38_plots
```

## HLA-DR+

### Naive PRE vs Naive POST

```{r}
naive_hladr_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(hladr_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15,  ylim = c(0, 50))
                               })
names(naive_hladr_plots) <- stims
naive_hladr_plots
```

### Conv PRE vs Conv POST

```{r}
conv_hladr_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(hladr_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+", 
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15,  ylim = c(0, 20))
                               })
names(conv_hladr_plots) <- stims
conv_hladr_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_hladr_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(hladr_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", groups_to_compare = c("Naive PRE", "Conv PRE"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(pre_hladr_plots) <- stims
pre_hladr_plots
```

### Naive POST vs Conv POST

```{r}
post_hladr_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(hladr_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "HLA-DR+",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(post_hladr_plots) <- stims
post_hladr_plots
```

## Central Memory

### Naive PRE vs Naive POST

```{r}
naive_cm_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cm_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Central Memory (CD45RA-CCR7+)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(naive_cm_plots) <- stims
naive_cm_plots
```

### Conv PRE vs Conv POST

```{r}
conv_cm_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cm_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"), 
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Central Memory (CD45RA-CCR7+)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(conv_cm_plots) <- stims
conv_cm_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_cm_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cm_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", groups_to_compare = c("Naive PRE", "Conv PRE"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Central Memory (CD45RA-CCR7+)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(pre_cm_plots) <- stims
pre_cm_plots
```

### Naive POST vs Conv POST

```{r}
post_cm_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cm_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Central Memory (CD45RA-CCR7+)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(post_cm_plots) <- stims
post_cm_plots
```

## Effector Memory

### Naive PRE vs Naive POST

```{r}
naive_em_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(em_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"),
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Effector Memory (CD45RA-CCR7-)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15,   ylim = c(0, 100))
                               })
names(naive_em_plots) <- stims
naive_em_plots
```

### Conv PRE vs Conv POST

```{r}
conv_em_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(em_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"), 
                                                paired = TRUE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Effector Memory (CD45RA-CCR7-)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15,  ylim = c(0, 65))
                               })
names(conv_em_plots) <- stims
conv_em_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_em_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(em_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", groups_to_compare = c("Naive PRE", "Conv PRE"), 
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Effector Memory (CD45RA-CCR7-)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(pre_em_plots) <- stims
pre_em_plots
```

### Naive POST vs Conv POST

```{r}
post_em_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(em_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"),
                                                paired = FALSE, fill_colors = fill_colors, group_by_colname = "Group", subtitle = "Effector Memory (CD45RA-CCR7-)",
                                                y_axis_text = "% CD4+ T Cells", y_axis_size = 15)
                               })
names(post_em_plots) <- stims
post_em_plots
```

# Save figures

```{r save boxplots pdf}
if(save_output) {
  cairo_pdf(file=here::here("out/post_compass_plots/activation_memory_plots/CD4_naive_hladr_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(naive_hladr_plots[["S1"]])
  print(naive_hladr_plots[["S2"]])
  dev.off()
  
  cairo_pdf(file=here::here("out/post_compass_plots/activation_memory_plots/CD4_naive_em_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica
  print(naive_em_plots[["S1"]])
  print(naive_em_plots[["S2"]])
  dev.off()
  
    cairo_pdf(file=here::here("out/post_compass_plots/activation_memory_plots/CD4_conv_cd38_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(conv_cd38_plots[["S1"]])
  print(conv_cd38_plots[["S2"]])
  dev.off()
  
    cairo_pdf(file=here::here("out/post_compass_plots/activation_memory_plots/CD4_conv_hladr_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(conv_hladr_plots[["S1"]])
  print(conv_hladr_plots[["S2"]])
  dev.off()
    
  cairo_pdf(file=here::here("out/post_compass_plots/activation_memory_plots/CD4_conv_em_plots.pdf"), width=4, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica
  print(conv_em_plots[["S1"]])
  print(conv_em_plots[["S2"]])
  dev.off()
}
```
