==> scripts/07_More_Frequency_Plots.Rmd <==
---
title: "More Frequency Plots"
author: "Jolie Phan"
date: "version `r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
editor_options:
  chunk_output_type: console
---

Plot frequencies of naive and convalescent CD4+ T cells which co-express IFNg and CD154.

Note: p-values are adjusted

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.width = 4, fig.height = 5)
```

# Load libraries

```{r, message = FALSE}
library(here)
library(flowWorkspace)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(ggbeeswarm)
library(cowplot)
library(ggh4x)
source(here::here("scripts/Helper_Functions.R")) # for make_mag_plots()
```

# Load data

```{r}
save_output <- TRUE

gs <- load_gs(here::here("out/GatingSets/HAARVIVAC_GatingSet_with_COMPASS_Subsets"))

if(!dir.exists(here::here("out/frequency_plots"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/frequency_plots")))
    dir.create(here::here("out/frequency_plots"), recursive = T)
}
```

```{r}
# Arial font setup. Downloaded afms from https://github.com/microsoft/microsoft-r-open/tree/ec3fd89e5fb5794bd8149905c134ad801bb61800
Arial <- Type1Font(family = "Arial",
                   metrics = c(here::here("data/Arial_afm/ArialMT.afm"),
                               here::here("data/Arial_afm/ArialMT-Bold.afm"),
                               here::here("data/Arial_afm/ArialMT-Italic.afm"),
                               here::here("data/Arial_afm/ArialMT-BoldItalic.afm")))
windowsFonts("Arial" = windowsFont("Arial"))
pdfFonts(Arial = Arial)
```

# Add boolean gate

## CD4 IFNg+CD154+

```{r}
# Get paths
cd4_ifng_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+"
cd4_cd154_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD154+"

# Add boolean gate
gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd4_ifng_path,
                                                         "&", cd4_cd154_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+", name = "IFNg+CD154+")
```

## CD4 IFNg+CD107a+

```{r}
# Get paths
cd4_ifng_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+"
cd4_cd107a_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/CD107a+" 

# Add boolean gate
gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd4_ifng_path,
                                                         "&", cd4_cd107a_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+", name = "IFNg+CD107a+")
```

## CD8 CD107a+HLADR|CD107a+CD38+

```{r}
# Get paths
cd8_cd107a_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD107a+"
cd3_hladr_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/HLADR+"
cd3_cd38_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD38+" 

# Add boolean gate
gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd8_cd107a_path,
                                                         "&", cd3_hladr_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+", name = "CD107a+HLADR+")

gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd8_cd107a_path,
                                                         "&", cd3_cd38_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+", name = "CD107a+CD38+")

cd8_cd107a_hladr_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD107a+HLADR+"
cd8_cd107a_cd38_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD107a+CD38+"

gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd8_cd107a_hladr_path,
                                                         "|", cd8_cd107a_cd38_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+", name = "CD107a+HLADR+|CD107a+CD38+")
```

## CD8 TEMRA CD107a+

```{r}
# Get paths
cd8_cd107a_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD107a+"
cd8_cd45ra_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD45RA+" 
cd8_ccr7_path <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CCR7+"

gs_pop_add(gs, eval(substitute(flowWorkspace::booleanFilter(v),
                               list(v = as.symbol(paste0("", cd8_cd107a_path,
                                                         "&", cd3_hladr_path,
                                                         "&!", cd8_ccr7_path))))),
           parent = "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+", name = "TEMRA_CD107a+")
```

Recompute the GatingSet

```{r}
flowWorkspace::recompute(gs)
```

# Extract frequencies

Exclude SEB.

## CD4 IFNg+CD154+

```{r}
cd4_ifng_cd154_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+CD154+"

ifng_cd154_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_ifng_cd154_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

ifng_cd154_counts$Group <- factor(ifng_cd154_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## CD4 IFNg+CD107a+

```{r}
cd4_ifng_cd107a_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD4+/IFNg+CD107a+"

ifng_cd107a_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd4_ifng_cd107a_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

ifng_cd107a_counts$Group <- factor(ifng_cd107a_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## CD8 CD107a+HLADR+|CD107a+CD38+

```{r}
cd8_cd107a_hladr_cd38_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/CD107a+HLADR+|CD107a+CD38+"

cd107a_hladr_cd38_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd8_cd107a_hladr_cd38_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

cd107a_hladr_cd38_counts$Group <- factor(cd107a_hladr_cd38_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

## CD8 TEMRA CD107a+

```{r}
cd8_temra_cd107a_parent <- "/Time/CD3+/CD14-CD19-/Lymphocytes/Singlet/Live/CD8+/TEMRA_CD107a+"

temra_cd107a_counts <- pData(gs) %>%
  tibble::rownames_to_column("rowname") %>%
  left_join(gs_pop_get_count_fast(gs, subpopulations = cd8_temra_cd107a_parent), by = c("rowname" = "name")) %>%
  dplyr::rename(Subpop = Count) %>%
  dplyr::select(rowname, "SAMPLE ID", "EXPERIMENT NAME", "PATIENT ID", Stim, Group, Timepoint, Subpop, ParentCount) %>%
  dplyr::filter(Stim != "SEB") %>%
  mutate(Freq = (Subpop/ParentCount)*100) 

temra_cd107a_counts$Group <- factor(temra_cd107a_counts$Group, levels = c("Naive PRE", "Naive POST", "Conv PRE", "Conv POST"))
```

# Plots

```{r}
stims <- c("DMSO", "S1", "S2", "NCAP") 

fill_colors <- list("Naive" = c("Naive PRE" = "#54A8E1", "Naive POST" = "#196293"),
                    "Conv" = c("Conv PRE" = "#EB4D4F", "Conv POST" = "#9D1213"),
                    "PRE" = c("Naive PRE" = "#54A8E1", "Conv PRE" = "#EB4D4F"),
                    "POST" = c("Naive POST" = "#196293", "Conv POST" = "#9D1213"))
```

## CD4 IFNg+CD154+

### Naive PRE vs Naive POST

```{r}
# There is an outlier in the S2-stimmed CD4 IFNg+CD154+ samples, the S2 data will be plotted separately w/o the outlier PTID 
naive_ifng_cd154_plots <- purrr::pmap(.l = list(stims[-3]),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd154_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive",
                                                groups_to_compare = c("Naive PRE", "Naive POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD154+"), y_axis_text = "% CD4+ T Cells",
                                                ylim = c(0, 1), y_axis_size = 20)
                               })
names(naive_ifng_cd154_plots) <- stims[-3]

ifng_cd154_counts_no_outlier <- ifng_cd154_counts %>%
  dplyr::filter(Stim == "S2") %>%
  dplyr::filter(`PATIENT ID` != "57H")
  
naive_s2_ifng_cd154_plot <- make_mag_plots(counts = ifng_cd154_counts, counts_no_outlier = ifng_cd154_counts_no_outlier, current_stim = "S2", num_comparisons = length(stims),
                                           compare_time = TRUE, keep = "Naive", groups_to_compare = c("Naive PRE", "Naive POST"), paired = TRUE, fill_colors = fill_colors,
                                           group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD154+"), y_axis_text = "% CD4+ T Cells", 
                                           ylim = c(0, 1), y_axis_size = 20)

naive_ifng_cd154_plots[["DMSO"]]
naive_ifng_cd154_plots[["S1"]]
naive_s2_ifng_cd154_plot
naive_ifng_cd154_plots[["NCAP"]]
```

### Conv PRE vs Conv POST

```{r}
conv_ifng_cd154_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd154_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv",
                                                groups_to_compare = c("Conv PRE", "Conv POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD154+"), y_axis_text = "% CD4+ T Cells",
                                                ylim = c(0, 1.75), y_axis_size = 20)
                               })
names(conv_ifng_cd154_plots) <- stims

conv_ifng_cd154_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_ifng_cd154_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd154_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", 
                                                groups_to_compare = c("Naive PRE", "Conv PRE"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD154+"), y_axis_text = "% CD4+ T Cells",
                                                y_axis_size = 20)
                               })
names(pre_ifng_cd154_plots) <- stims

pre_ifng_cd154_plots
```

### Naive POST vs Conv POST

```{r}
# Drop DMSO
post_ifng_cd154_plots <- purrr::pmap(.l = list(stims[-1]),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd154_counts, current_stim = n, num_comparisons = 3, compare_time = FALSE, keep = "POST", 
                                                groups_to_compare = c("Naive POST", "Conv POST"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD154+"), y_axis_text = "% CD4+ T Cells",
                                                ylim = c(0, 2.75), y_axis_size = 20)
                               })
names(post_ifng_cd154_plots) <- stims[-1]

post_ifng_cd154_plots
```

## CD4 IFNg+CD107a+

### Naive PRE vs Naive POST

```{r}
naive_ifng_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive",
                                                groups_to_compare = c("Naive PRE", "Naive POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells",
                                                ylim = c(0, 0.06), y_axis_size = 20)
                               })
names(naive_ifng_cd107a_plots) <- stims

naive_ifng_cd107a_plots
```

### Conv PRE vs Conv POST

```{r}
# Drop DMSO and NCAP 
# There is an outlier in the S2-stimmed CD4 IFNg+CD107a+ samples, the S2 data will be plotted separately w/o the outlier PTID 
conv_s1_ifng_cd107a_plot <- make_mag_plots(counts = ifng_cd107a_counts, current_stim = "S1", num_comparisons = 2,
                                           compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"), paired = TRUE, fill_colors = fill_colors,
                                           group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells", 
                                           ylim = c(0, 0.35), y_axis_size = 20)

ifng_cd107a_counts_no_outlier <- ifng_cd107a_counts %>%
  dplyr::filter(Stim == "S2") %>%
  dplyr::filter(`PATIENT ID` != "242C")
  
conv_s2_ifng_cd107a_plot <- make_mag_plots(counts = ifng_cd107a_counts, counts_no_outlier = ifng_cd107a_counts_no_outlier, current_stim = "S2", num_comparisons = 2,
                                           compare_time = TRUE, keep = "Conv", groups_to_compare = c("Conv PRE", "Conv POST"), paired = TRUE, fill_colors = fill_colors,
                                           group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells", 
                                           ylim = c(0, 0.35), y_axis_size = 20)

conv_s1_ifng_cd107a_plot
conv_s2_ifng_cd107a_plot
```

### Naive PRE vs Conv PRE

```{r}
pre_ifng_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(ifng_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", 
                                                groups_to_compare = c("Naive PRE", "Conv PRE"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells",
                                                y_axis_size = 20)
                               })
names(pre_ifng_cd107a_plots) <- stims

pre_ifng_cd107a_plots
```

### Naive POST vs Conv POST

```{r}
# Drop DMSO and NCAP 
# There is an outlier in the S1-stimmed CD4 IFNg+CD107a+ samples, the S1 data will be plotted separately w/o the outlier PTID 
post_s2_ifng_cd107a_plot <- make_mag_plots(counts = ifng_cd107a_counts, current_stim = "S2", num_comparisons = 2,
                                           compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"), paired = FALSE, fill_colors = fill_colors,
                                           group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells", 
                                           ylim = c(0, 0.17), y_axis_size = 20)

ifng_cd107a_counts_no_outlier <- ifng_cd107a_counts %>%
  dplyr::filter(Stim == "S1") %>%
  dplyr::filter(`PATIENT ID` != "12C")
  
post_s1_ifng_cd107a_plot <- make_mag_plots(counts = ifng_cd107a_counts, counts_no_outlier = ifng_cd107a_counts_no_outlier, current_stim = "S1", num_comparisons = 2,
                                           compare_time = FALSE, keep = "POST", groups_to_compare = c("Naive POST", "Conv POST"), paired = FALSE, fill_colors = fill_colors,
                                           group_by_colname = "Group", subtitle = paste0("IFN", "\U03B3", "+CD107a+"), y_axis_text = "% CD4+ T Cells", 
                                           ylim = c(0, 0.17), y_axis_size = 20)


post_s1_ifng_cd107a_plot
post_s2_ifng_cd107a_plot
```

## CD8 CD107a+HLADR+CD38+

### Naive PRE vs Naive POST

```{r}
# There is nothing in DMSO, so drop it here
naive_cd107a_hladr_cd38_plots <- purrr::pmap(.l = list(stims[-1]),
                               .f = function(n) {
                                 make_mag_plots(cd107a_hladr_cd38_counts, current_stim = n, num_comparisons = 3, compare_time = TRUE, keep = "Naive",
                                                groups_to_compare = c("Naive PRE", "Naive POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = "CD107a+HLA-DR+ or CD107a+CD38+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })

names(naive_cd107a_hladr_cd38_plots) <- stims[-1]

naive_cd107a_hladr_cd38_plots
```

### Conv PRE vs Conv POST

```{r}
conv_cd107a_hladr_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd107a_hladr_cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv",
                                                groups_to_compare = c("Conv PRE", "Conv POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = "CD107a+HLA-DR+ or CD107a+CD38+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(conv_cd107a_hladr_cd38_plots) <- stims

conv_cd107a_hladr_cd38_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_cd107a_hladr_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd107a_hladr_cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", 
                                                groups_to_compare = c("Naive PRE", "Conv PRE"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = "CD107a+HLA-DR+ or CD107a+CD38+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(pre_cd107a_hladr_cd38_plots) <- stims

pre_cd107a_hladr_cd38_plots
```

### Naive POST vs Conv POST

```{r}
post_cd107a_hladr_cd38_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(cd107a_hladr_cd38_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", 
                                                groups_to_compare = c("Naive POST", "Conv POST"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = "CD107a+HLA-DR+ or CD107a+CD38+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(post_cd107a_hladr_cd38_plots) <- stims

post_cd107a_hladr_cd38_plots
```

## CD8 TEMRA CD107a+

### Naive PRE vs Naive POST

```{r}
naive_temra_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(temra_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Naive",
                                                groups_to_compare = c("Naive PRE", "Naive POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = "TEMRA CD107a+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })

names(naive_temra_cd107a_plots) <- stims

naive_temra_cd107a_plots
```

### Conv PRE vs Conv POST

```{r}
conv_temra_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(temra_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = TRUE, keep = "Conv",
                                                groups_to_compare = c("Conv PRE", "Conv POST"), paired = TRUE, fill_colors = fill_colors,
                                                group_by_colname = "Group", subtitle = "TEMRA CD107a+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(conv_temra_cd107a_plots) <- stims

conv_temra_cd107a_plots
```

### Naive PRE vs Conv PRE

```{r}
pre_temra_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(temra_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "PRE", 
                                                groups_to_compare = c("Naive PRE", "Conv PRE"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = "TEMRA CD107a+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(pre_temra_cd107a_plots) <- stims

pre_temra_cd107a_plots
```

### Naive POST vs Conv POST

```{r}
post_temra_cd107a_plots <- purrr::pmap(.l = list(stims),
                               .f = function(n) {
                                 make_mag_plots(temra_cd107a_counts, current_stim = n, num_comparisons = length(stims), compare_time = FALSE, keep = "POST", 
                                                groups_to_compare = c("Naive POST", "Conv POST"), paired = FALSE, fill_colors = fill_colors, 
                                                group_by_colname = "Group", subtitle = "TEMRA CD107a+", y_axis_text = "% CD8+ T Cells",
                                                y_axis_size = 20)
                               })
names(post_temra_cd107a_plots) <- stims

post_temra_cd107a_plots
```

# Save figures

```{r save boxplots pdf}
if(save_output) {
  cairo_pdf(file=here::here("out/frequency_plots/CD4_naive_ifng_cd154_plot.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(naive_ifng_cd154_plots[["DMSO"]])
  print(naive_ifng_cd154_plots[["S1"]])
  print(naive_s2_ifng_cd154_plot)
  print(naive_ifng_cd154_plots[["NCAP"]])
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/CD4_conv_ifng_cd154_plot.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(conv_ifng_cd154_plots[["DMSO"]])
  print(conv_ifng_cd154_plots[["S1"]])
  print(conv_ifng_cd154_plots[["S2"]])
  print(conv_ifng_cd154_plots[["NCAP"]])
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/CD4_post_ifng_cd154_plot.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(post_ifng_cd154_plots[["S1"]])
  print(post_ifng_cd154_plots[["S2"]])
  print(post_ifng_cd154_plots[["NCAP"]])
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/CD4_conv_ifng_cd107a_plot.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(conv_s1_ifng_cd107a_plot)
  print(conv_s2_ifng_cd107a_plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/CD4_post_ifng_cd107a_plot.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(post_s1_ifng_cd107a_plot)
  print(post_s2_ifng_cd107a_plot)
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/S1_CD8_temra_cd107a_plots.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(naive_temra_cd107a_plots[["S1"]])
  print(conv_temra_cd107a_plots[["S1"]])
  print(post_temra_cd107a_plots[["S1"]])
  dev.off()
  
  cairo_pdf(file=here::here("out/frequency_plots/S2_CD8_cd107a_hladr_or_cd38_plots.pdf"), width=4, height=5,
            onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(naive_cd107a_hladr_cd38_plots[["S2"]])
  print(conv_cd107a_hladr_cd38_plots[["S2"]])
  print(post_cd107a_hladr_cd38_plots[["S2"]])
  dev.off()
}
```
